backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
PyJWT
python-multipart
psycopg2-binary
passlib[bcrypt]
pytest
flake8
httpx
slowapi
